Updates 19-4-2023
Module sysdata.rda, is updated with the list of soil grid names (bdod_0-5cm_mean, it was BLDFIE_M_sl1)

GetSG.R it is updated the code, by changing the url (https://www.isric.org/)

SoilParams.R it is updated: are changed the soil grid names, sl1,sl2 ... is substituted with sd1, sd2, ... (check the info at https://www.isric.org/explore/soilgrids/faq-soilgrids-2017#What_is_SoilGrids).
due to a problem with tiles it is updated the a. a=1.